Add libfuse2 package to vm.yml and rearrange commands in initial-install.sh
In vm.yml, the libfuse2 package was added to the list of packages. In initial-install.sh, the order of the commands was rearranged to ensure that the necessary permissions are granted before certain actions are performed. Furthermore, 'sudo' was added to the removal command for the jetbrains toolbox file to avoid possible permission issues.